refactor: improve translation file handling
Refactor translation file handling by extracting related
types into a separate module file. Replace `getFiles` with
`getTranslationFiles` in tests to reflect this change,
improving code clarity and separation of concerns.
