+++ src/libs/store/store-file-location-store/src/lib/file-location-store.types.ts
@@ -1,0 +1,22 @@
+import { Uri } from 'vscode';
+
+import { FileType } from '@i18n-weave/util/util-enums';
+
+type Metadata = {
+  uri: Uri;
+  lastModified: Date;
+  type: FileType;
+};
+
+type FileData = {
+  content: string;
+  metaData: Metadata;
+};
+
+export type TranslationFile = {
+  language: string;
+  namespace: string;
+  dialect?: string;
+} & FileData;
+
+export type CodeFile = FileData;
